pyproject.toml
[project]
name = "nimbly-api"
version = "0.1.0"
description = "a RESTful API that allows users to manage EC2 instances."
authors = [
    {name = "Rhythm Garg",email = "rhythmgarg05@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.1,<0.35.0)",
    "boto3[crt] (>=1.37.33,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "jwt (>=1.3.1,<2.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "motor (>=3.7.0,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
boto3-stubs = {extras = ["ec2", "s3", "sts"], version = "^1.37.33"}

